filter by election event in controller, break up models module
