+++ censusreporter/api/controller/elections.py
@@ -26,7 +26,7 @@ def make_party_acronym(name):
         return acronym
 
 
-def get_elections_profile(geo_code, geo_level):
+def get_elections_profile(geo_code, geo_level, election="municipal 2011"):
     session = get_session()
 
     try:
@@ -40,8 +40,9 @@ def get_elections_profile(geo_code, geo_level):
         # only consider proportional representative (PR) votes for now
         votes_by_party = session \
                 .query(Votes.party, func.sum(Votes.valid_votes).label('vote_sum')) \
-                .filter(getattr(Votes, field_name) == geo_code,
-                        Votes.ballot_type == 'PR') \
+                .filter(getattr(Votes, field_name) == geo_code) \
+                .filter(Votes.ballot_type == 'PR') \
+                .filter(Votes.electoral_event == election) \
                 .group_by(Votes.party) \
                 .having(func.sum(Votes.valid_votes) > 0) \
                 .order_by(desc('vote_sum')) \
@@ -49,8 +50,9 @@ def get_elections_profile(geo_code, geo_level):
         subquery = session \
                 .query(Votes.total_votes, Votes.registered_voters,
                        Votes.voter_turnout, Votes.spoilt_votes) \
-                .filter(getattr(Votes, field_name) == geo_code,
-                        Votes.ballot_type == 'PR') \
+                .filter(getattr(Votes, field_name) == geo_code) \
+                .filter(Votes.ballot_type == 'PR') \
+                .filter(Votes.electoral_event == election) \
                 .distinct(Votes.voting_district_code) \
                 .subquery() 
         total_votes, spoilt_votes, registered_voters, average_turnout = session \
@@ -90,20 +92,22 @@ def get_elections_profile(geo_code, geo_level):
             'universe': 'Proportional representative (PR) ballots only'
         }
 
-        return [{
-            'electoral_event': '2011 municipal elections',
-            'party_distribution': party_data,
-            'registered_voters': {
-                "name": "Number of registered voters",
-                "values": {"this": registered_voters},
-                "error": {"this": 0.0},
-            },
-            'average_turnout': {
-                "name": "Of registered voters cast their vote",
-                "values": {"this": round(average_turnout, 2)},
-                "error": {"this": 0.0},
+        return {
+            election.replace(' ', '_'): {
+                'name': election, 
+                'party_distribution': party_data,
+                'registered_voters': {
+                    "name": "Number of registered voters",
+                    "values": {"this": registered_voters},
+                    "error": {"this": 0.0},
+                },
+                'average_turnout': {
+                    "name": "Of registered voters cast their vote",
+                    "values": {"this": round(average_turnout, 2)},
+                    "error": {"this": 0.0},
+                }
             }
-        }]
+        }
 
     except KeyError:
         raise ValueError('Invalid geo_level: %s' % geo_level)
