+++ censusreporter/api/models/__init__.py
@@ -1,0 +1,7 @@
+from .base import Base, Ward, Municipality, District, Province
+from .census import get_model_from_fields
+from .elections import Votes
+
+
+__all__ = ['Base', 'Ward', 'Municipality', 'District', 'Province',
+           'get_model_from_fields', 'Votes']
