+++ censusreporter/apps/census/templates/profile/_blocks/_stat_list.html
@@ -20,9 +20,6 @@
 			{% endif %}
         </span>
         <span class="name">{{ stat.name }}</span>
-        {% if stat.metadata.acs_release != geography.census_release %}
-        	<span class="small">* {{ stat.metadata.acs_release }} data</span>
-        {% endif %}
 		
 		{% if stat_class == 'secondary' %}
 			<span class="context">
